implemented stack to solve max_depth_parentheses
guaranteed that the expression is valid, so we do not need to worry about the case in which we pop an empy stack.

push if we see a front bracket pop if we see a back bracket.

For every iteration we check if the current length of the stack is greater than the max count, and replace the max count if it is.

Run time -> O(n)

Space time -> O(n)